组织管理模块的权限验证处理完毕 添加一个组织关系实体用于组织关系扩展

+++ AuthorizationCenter/Controllers/OrganizationController.cs
@@ -25,12 +25,17 @@ namespace AuthorizationCenter.Controllers
         /// <summary>
         /// 管理
         /// </summary>
-        public IOrganizationManager<OrganizationJson> OrganizationManager { get; set; }
+        IOrganizationManager<OrganizationJson> OrganizationManager { get; set; }
+
+        /// <summary>
+        /// 角色组织权限管理
+        /// </summary>
+        IRoleOrgPerManager<RoleOrgPerJson> RoleOrgPerManager { get; set; }
 
         /// <summary>
         /// 日志记录器
         /// </summary>
-        public ILogger Logger = LoggerManager.GetLogger(nameof(OrganizationController));
+        readonly ILogger Logger = LoggerManager.GetLogger<OrganizationController>();
 
         /// <summary>
         /// 组织管理
@@ -42,7 +47,8 @@ namespace AuthorizationCenter.Controllers
         }
 
         /// <summary>
-        /// 列表 -MVC
+        /// [MVC] 组织管理-组织列表
+        /// 只能查看ORG_QUERY的组织森林
         /// </summary>
         /// <returns></returns>
         // GET: Organization
@@ -50,7 +56,8 @@ namespace AuthorizationCenter.Controllers
         {
             try
             {
-                // TODO 筛选出登陆用户可见的组织 -根据用户找到角色 -根据角色找到组织 -根据父组织找到所有子组织
+                // 1. 权限验证 ORG_QUERY U.ID-[UR]->R.ID|P.ID-[ROP]->Count(O.ID)>0
+                // 2. 业务处理 -查询有权组织 森林列表
                 var orgnazitions = await OrganizationManager.FindByUserId(SignUser.Id);
 
                 Logger.Trace($"[{nameof(Index)}] 响应数据：\r\n{JsonUtil.ToJson(orgnazitions)}");
@@ -61,26 +68,28 @@ namespace AuthorizationCenter.Controllers
             }
             catch (Exception e)
             {
-                Logger.Error($"[{nameof(Index)}] 组织列表获取失败：\r\n" + e);
+                ModelState.AddModelError("All", "组织列表获取失败");
+                Logger.Error($"[{nameof(Index)}] 服务器错误：\r\n" + e);
                 return View();
             }
         }
 
         /// <summary>
-        /// 列表 -API -只能看到自己所在的组织（查找方式，通过角色->三者关联->组织）
+        /// [API] 列表 -代码编写中 
+        /// -只能看到自己所在的组织（查找方式，通过角色->三者关联->组织）
         /// </summary>
         /// <returns></returns>
         public async Task<PageResponesBody<OrganizationJson>> List()
         {
             try
             {
-                var orgnazitions = await OrganizationManager.FindByUserId(SignUser.Id);
+                var organizations = await OrganizationManager.FindByUserId(SignUser.Id);
 
-                Console.WriteLine(JsonUtil.ToJson(orgnazitions));
+                Console.WriteLine(JsonUtil.ToJson(organizations));
 
                 return new PageResponesBody<OrganizationJson>
                 {
-                    Data = orgnazitions.ToList()
+                    Data = organizations.ToList()
                 };
             }
             catch (Exception e)
@@ -105,17 +114,25 @@ namespace AuthorizationCenter.Controllers
             }
 
             try{
+                // 1. 权限检查
+                if(!await RoleOrgPerManager.HasPermission(SignUser.Id, id, Constants.ORG_QUERY))
+                {
+                    Logger.Warn($"[{nameof(Details)}] 权限不足 用户[{SignUser.SignName}]({SignUser.Id})没有权限({Constants.ORG_QUERY})操作组织({id})");
+                    ModelState.AddModelError("All", "权限不足");
+                    return View(nameof(Index));
+                }
+                // 2. 业务处理
                 var organization = await OrganizationManager.FindById(id).SingleOrDefaultAsync();
+                Logger.Trace($"[{nameof(Details)}] 响应数据:\r\n{JsonUtil.ToJson(organization)}");
                 if (organization == null)
                 {
                     return NotFound();
                 }
-
                 return View(organization);
             }
             catch (Exception e)
             {
-                Logger.Error($"[{nameof(Details)}] 组织详情查看失败" + e);
+                Logger.Error($"[{nameof(Details)}] 服务器错误:\r\n{e}");
                 return View(nameof(Index));
             }
         }
@@ -123,6 +140,7 @@ namespace AuthorizationCenter.Controllers
         /// <summary>
         /// 创建 -MVC -组织列表（用来选择父组织）
         /// </summary>
+        /// <param name="id">父组织ID</param>
         /// <returns></returns>
         // GET: Organization/Create
         [TypeFilter(typeof(CheckPermission), Arguments = new object[] { Constants.ORG_MANAGE })]
@@ -131,24 +149,35 @@ namespace AuthorizationCenter.Controllers
             Logger.Trace($"[{nameof(Create)}] 请求参数: id: " +id);
             try
             {
+                // 1. 权限检查
+                if (!await RoleOrgPerManager.HasPermission(SignUser.Id, id, Constants.ORG_CREATE))
+                {
+                    Logger.Warn($"[{nameof(Details)}] 权限不足 用户[{SignUser.SignName}]({SignUser.Id})没有权限({Constants.ORG_CREATE})操作组织({id})");
+                    ModelState.AddModelError("All", "权限不足");
+                    return View(nameof(Index));
+                }
+                // 2. 业务处理
                 var organizations = await OrganizationManager.Find().ToListAsync();
+                Logger.Trace($"[{nameof(Index)}] 响应数据：\r\n{JsonUtil.ToJson(organizations)}");
                 if (id != null)
                 {
                     organizations =organizations.Where(org => org.Id == id).ToList();
                 }
                 ViewData["OrgId"] = new SelectList(organizations, nameof(Organization.Id), nameof(Organization.Name), id);
+                return View();
             }
             catch (Exception e)
             {
-                Logger.Error($"[{nameof(Create)}] 组织列表查询失败：\r\n" + e);
+                Logger.Error($"[{nameof(Create)}] 服务器错误：\r\n" + e);
+                ModelState.AddModelError("All", "页面跳转失败");
+                return View(nameof(Index));
             }
-            return View();
         }
 
         /// <summary>
         /// 创建 -MVC
         /// </summary>
-        /// <param name="organization"></param>
+        /// <param name="organization">组织</param>
         /// <returns></returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -157,6 +186,20 @@ namespace AuthorizationCenter.Controllers
             Logger.Trace($"[{nameof(Create)}] 新增：请求参数：\r\n" + JsonUtil.ToJson(organization));
             try
             {
+                // 0. 参数检查
+                if (organization == null || organization.ParentId == null)
+                {
+                    ModelState.AddModelError("All", "参数异常");
+                    return View();
+                }
+                // 1. 权限检查
+                if (!await RoleOrgPerManager.HasPermission(SignUser.Id, organization.ParentId, Constants.ORG_CREATE))
+                {
+                    Logger.Warn($"[{nameof(Details)}] 权限不足 用户[{SignUser.SignName}]({SignUser.Id})没有权限({Constants.ORG_CREATE})操作组织({organization.ParentId})");
+                    ModelState.AddModelError("All", "权限不足");
+                    return View(nameof(Index));
+                }
+                // 2. 业务处理
                 // 检查是否存在循环（允许有向无环图的产生，而不仅仅是树形结构）
                 await OrganizationManager.Create(organization);
             }
@@ -176,27 +219,37 @@ namespace AuthorizationCenter.Controllers
         // GET: Organization/Edit/5
         public async Task<IActionResult> Edit(string id)
         {
+            Logger.Trace($"[{nameof(Edit)}] 请求参数: id: " + id);
+            // 0. 参数检查
             if (id == null)
             {
                 return NotFound();
             }
-            Logger.Trace($"[{nameof(Edit)}] 请求参数: id: " + id);
-            OrganizationJson organization = null;
             try
             {
-                organization = await OrganizationManager.FindById(id).SingleOrDefaultAsync();
+                // 1. 权限检查
+                if (!await RoleOrgPerManager.HasPermission(SignUser.Id, id, Constants.ORG_UPDATE))
+                {
+                    Logger.Warn($"[{nameof(Details)}] 权限不足 用户[{SignUser.SignName}]({SignUser.Id})没有权限({Constants.ORG_UPDATE})操作组织({id})");
+                    ModelState.AddModelError("All", "权限不足");
+                    return View(nameof(Index));
+                }
+                // 2. 业务处理
+                var organization = await OrganizationManager.FindById(id).SingleOrDefaultAsync();
+
+                if (organization == null)
+                {
+                    return NotFound();
+                }
+                Logger.Trace($"[{nameof(Edit)}] 响应数据：\r\n" + JsonUtil.ToJson(organization));
+                return View(organization);
             }
             catch(Exception e)
             {
                 Logger.Error($"[{nameof(Edit)}] 服务器错误：\r\n" + e);
+                ModelState.AddModelError("All", "编辑界面跳转失败");
+                return View(nameof(Index));
             }
-
-            if (organization == null)
-            {
-                return NotFound();
-            }
-            Logger.Trace($"[{nameof(Edit)}] 响应体：\r\n" + JsonUtil.ToJson(organization));
-            return View(organization);
         }
 
         /// <summary>
@@ -206,66 +259,81 @@ namespace AuthorizationCenter.Controllers
         /// <param name="organization"></param>
         /// <returns></returns>
         // POST: Organization/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(string id, [Bind("Id,ParentId,Name,Description")] OrganizationJson organization)
         {
+            Logger.Trace($"[{nameof(Edit)}] 请求参数: id: {id}\r\n{JsonUtil.ToJson(organization)}");
+            // 0. 参数检查
             if (id != organization.Id)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            try
             {
-                try
+                // 1. 权限检查
+                if (!await RoleOrgPerManager.HasPermission(SignUser.Id, id, Constants.ORG_UPDATE))
                 {
-                    await OrganizationManager.Update(organization);
+                    Logger.Warn($"[{nameof(Details)}] 权限不足 用户[{SignUser.SignName}]({SignUser.Id})没有权限({Constants.ORG_UPDATE})操作组织({id})");
+                    ModelState.AddModelError("All", "权限不足");
+                    return View(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!await OrganizationManager.Exist(org => org.Id == id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                // 2. 业务处理
+                await OrganizationManager.Update(organization);
+                return View(organization);
             }
-            return View(organization);
+            catch (Exception e)
+            {
+                Logger.Error($"[{nameof(Edit)}] 服务器错误:\r\n{e}");
+                ModelState.AddModelError("All", "保存失败");
+                return View(nameof(Index));
+            }
         }
 
         /// <summary>
-        /// 删除 -MVC
+        /// [MVC] 组织管理-组织删除界面跳转
         /// </summary>
-        /// <param name="id"></param>
+        /// <param name="id">组织ID</param>
         /// <returns></returns>
         // GET: Organization/Delete/5
         [TypeFilter(typeof(CheckPermission))]
         public async Task<IActionResult> Delete(string id)
         {
+            // 0. 参数检查
             if (id == null)
             {
                 return NotFound();
             }
-
-            var organization = await OrganizationManager.FindById(id).SingleOrDefaultAsync();
-            if (organization == null)
+            try
             {
-                return NotFound();
+                // 1. 权限检查
+                if (!await RoleOrgPerManager.HasPermission(SignUser.Id, id, Constants.ORG_DELETE))
+                {
+                    Logger.Warn($"[{nameof(Details)}] 权限不足 用户[{SignUser.SignName}]({SignUser.Id})没有权限({Constants.ORG_DELETE})操作组织({id})");
+                    ModelState.AddModelError("All", "权限不足");
+                    return View(nameof(Index));
+                }
+                // 2. 业务处理
+                var organization = await OrganizationManager.FindById(id).SingleOrDefaultAsync();
+                Logger.Trace($"[{nameof(Edit)}] 响应数据：\r\n" + JsonUtil.ToJson(organization));
+                if (organization == null)
+                {
+                    return NotFound();
+                }
+                return View(organization);
             }
-
-            return View(organization);
+            catch (Exception e)
+            {
+                Logger.Error($"[{nameof(Edit)}] 服务器错误:\r\n{e}");
+                ModelState.AddModelError("All", "界面跳转失败");
+                return View(nameof(Index));
+            }
         }
 
         /// <summary>
-        /// 通过ID删除 -API
+        /// [API] 组织管理-通过ID删除 -代码编写中
         /// </summary>
-        /// <param name="id"></param>
+        /// <param name="id">组织ID</param>
         /// <returns></returns>
         //[HttpDelete]
         [TypeFilter(typeof(CheckPermission), Arguments = new object[]{Constants.ORG_MANAGE})]
@@ -296,26 +364,46 @@ namespace AuthorizationCenter.Controllers
         }
 
         /// <summary>
-        /// 删除确认 -MVC
+        /// [MVC] 组织管理-组织删除确认
         /// </summary>
-        /// <param name="id"></param>
+        /// <param name="id">组织ID</param>
         /// <returns></returns>
         // POST: Organization/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
-            await OrganizationManager.DeleteById(id);
-            return RedirectToAction(nameof(Index));
+
+            // 0. 参数检查
+            if(id == null)
+            {
+                return NotFound();
+            }
+            try
+            {
+                // 1. 权限检查
+                if (!await RoleOrgPerManager.HasPermission(SignUser.Id, id, Constants.ORG_DELETE))
+                {
+                    Logger.Warn($"[{nameof(Details)}] 权限不足 用户[{SignUser.SignName}]({SignUser.Id})没有权限({Constants.ORG_DELETE})操作组织({id})");
+                    ModelState.AddModelError("All", "权限不足");
+                    return View(nameof(Index));
+                }
+                // 2. 业务处理
+                await OrganizationManager.DeleteById(id);
+                return View(nameof(Index));
+            }
+            catch(Exception e)
+            {
+                Logger.Error($"[{nameof(Edit)}] 服务器错误:\r\n{e}");
+                ModelState.AddModelError("All", "保存失败");
+                return View(nameof(Delete), id);
+            }
         }
 
         /// <summary>
         /// 获取登陆用户简要信息 -每次都是新建一个UserBaseJson对象
         /// </summary>
         /// <returns></returns>
-        /// <summary>
-        /// 登陆用户
-        /// </summary>
         private UserJson SignUser
         {
             get
